pyproject.toml
[tool.poetry]
name = "pullama"
version = "0.1.1"
description = "A git diff Summarizer backed by LLM on Ollama"
authors = ["Paulo Suzart <paulo.suzart@getmoss.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
gitpython = "^3.1.40"
click = "^8.1.7"
langchain = "^0.0.350"
unstructured = "^0.11.2"
esprima = "^4.0.1"
fastembed = "^0.1.1"
qdrant-client = "^1.7.0"
sync = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
